IBX-11974: Moved DeduplicateStamp to contracts and deprecated bundle class
