+++ src/bundle/Stamp/DeduplicateStamp.php
@@ -8,52 +8,26 @@ declare(strict_types=1);
 
 namespace Ibexa\Bundle\Messenger\Stamp;
 
-use Symfony\Component\Lock\Key;
-use Symfony\Component\Messenger\Exception\LogicException;
-use Symfony\Component\Messenger\Stamp\StampInterface;
+use Ibexa\Contracts\Messenger\Stamp\DeduplicateStamp as ContractsDeduplicateStamp;
 
 /**
- * (c) Fabien Potencier <fabien@symfony.com>.
- *
- * Original code: https://github.com/symfony/symfony/blob/7.3/src/Symfony/Component/Messenger/Stamp/DeduplicateStamp.php
+ * @deprecated since Ibexa 5.0.9, use {@see ContractsDeduplicateStamp} instead.
  */
-final class DeduplicateStamp implements StampInterface
+class DeduplicateStamp extends ContractsDeduplicateStamp
 {
-    private Key $key;
-
-    private ?float $ttl;
-
-    private bool $onlyDeduplicateInQueue;
-
     public function __construct(
         string $key,
         ?float $ttl = 300.0,
         bool $onlyDeduplicateInQueue = false
     ) {
-        if (!class_exists(Key::class)) {
-            throw new LogicException(sprintf(
-                'You cannot use the "%s" as the Lock component is not installed. Try running "composer require symfony/lock".',
-                self::class,
-            ));
-        }
+        trigger_deprecation(
+            'ibexa/messenger',
+            '5.0.9',
+            'The "%s" class is deprecated, use "%s" instead.',
+            self::class,
+            ContractsDeduplicateStamp::class,
+        );
 
-        $this->key = new Key($key);
-        $this->ttl = $ttl;
-        $this->onlyDeduplicateInQueue = $onlyDeduplicateInQueue;
-    }
-
-    public function onlyDeduplicateInQueue(): bool
-    {
-        return $this->onlyDeduplicateInQueue;
-    }
-
-    public function getKey(): Key
-    {
-        return $this->key;
-    }
-
-    public function getTtl(): ?float
-    {
-        return $this->ttl;
+        parent::__construct($key, $ttl, $onlyDeduplicateInQueue);
     }
 }
